third_party/WebKit/LayoutTests/vr/getVRDisplays_detached.html

<!DOCTYPE html>
<html>
<head>
<title>Detached use of navigator.getVRDisplays()</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<iframe sandbox="allow-same-origin" id="subframe"></iframe>
<script>
test(() => {
    var nav = window.frames[0].navigator;
    document.getElementById("subframe").remove();
    // A promise won't be rejected when in a detached context,
    // so the undefined value is returned.
    assert_equals(nav.getVRDisplays(), undefined);
});
</script>
</body>
</html>
